Rename ewts to topoflow_glacier_ewts
