+++ pyproject.toml
@@ -12,7 +12,7 @@ allow-direct-references = true
 [tool.hatch.build.targets.wheel]
 packages = [
   "src/topoflow_glacier",
-  "src/ewts",
+  "src/topoflow_glacier_ewts",
   "tools"
 ]
 
@@ -43,7 +43,7 @@ dependencies = [
   "pyyaml==6.0.3",
   "timezonefinder>=8.0.0",
   "gs==1.1.1",
-  "ewts @ file:src/ewts",
+  "topoflow_glacier_ewts @ file:src/topoflow_glacier_ewts",
 ]
 
 [project.optional-dependencies]
