+++ src/topoflow_glacier/physics/context.py
@@ -4,7 +4,7 @@ from collections.abc import Iterable, Iterator
 
 import numpy as np
 from pydantic import BaseModel, ConfigDict
-from ewts import MODULE_NAME
+from topoflow_glacier_ewts import MODULE_NAME
 
 import logging
 LOG = logging.getLogger(MODULE_NAME)
