+++ src/topoflow_glacier_ewts/pyproject.toml
@@ -3,6 +3,6 @@ requires = ["setuptools>=61"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "ewts"
+name = "topoflow-glacier-ewts"
 version = "0.1.0"
 requires-python = ">=3.6"
