+++ tests/topoflow_glacier_ewts/test_paths.py
@@ -2,9 +2,9 @@ import os
 import getpass
 from datetime import datetime
 import pytest
-from ewts import paths
-from ewts.paths import create_timestamp, get_log_file_path
-from ewts.constants import MODULE_NAME, EV_MODULE_LOGFILEPATH, EV_NGEN_LOGFILEPATH
+from topoflow_glacier_ewts import paths
+from topoflow_glacier_ewts.paths import create_timestamp, get_log_file_path
+from topoflow_glacier_ewts.constants import MODULE_NAME, EV_MODULE_LOGFILEPATH, EV_NGEN_LOGFILEPATH
 
 # -------------------------------
 # Fixture for a clean log environment
